+++ tests/e2e/e2e_suite_test.go
@@ -5,12 +5,10 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"regexp"
 	"strings"
 	"testing"
 
 	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 	"github.com/evmos/evmos/v16/app"
 	"github.com/evmos/evmos/v16/encoding"
@@ -339,16 +337,6 @@ func (s *IntegrationTestSuite) executeQueries() {
 	s.T().Logf("executed all queries successfully")
 }
 
-// executeTransactions executes some sample transactions to check they are still working after the upgrade.
-func (s *IntegrationTestSuite) executeTransactions() {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	chainID := utils.TestnetChainID + "-1"
-
-	s.sendBankTransfer(ctx, chainID)
-}
-
 // TearDownSuite kills the running container, removes the network and mount path
 func (s *IntegrationTestSuite) TearDownSuite() {
 	if s.upgradeParams.SkipCleanup {
@@ -366,45 +354,4 @@ func (s *IntegrationTestSuite) TearDownSuite() {
 
 	s.T().Log("removing mount path...")
 	s.Require().NoError(os.RemoveAll(strings.Split(s.upgradeParams.MountPath, ":")[0]), "can't remove mount path")
-}
-
-// SendBankTransfer sends a bank transfer to check that the transactions are still working after the upgrade.
-func (s *IntegrationTestSuite) sendBankTransfer(ctx context.Context, chainID string) {
-	receiver := "evmos1jcltmuhplrdcwp7stlr4hlhlhgd4htqh3a79sq"
-	sentCoins := sdk.Coins{sdk.NewInt64Coin("aevmos", 10000000000)}
-
-	balancePre, err := s.upgradeManager.GetBalance(ctx, s.upgradeParams.ChainID, receiver)
-	s.Require().NoError(err, "can't get balance of receiver account")
-
-	// send some tokens between accounts to check transactions are still working
-	exec, err := s.upgradeManager.CreateModuleTxExec(upgrade.E2ETxArgs{
-		ModuleName: "bank",
-		SubCommand: "send",
-		Args:       []string{"mykey", receiver, sentCoins.String()},
-		ChainID:    chainID,
-		From:       "mykey",
-	})
-	s.Require().NoError(err, "failed to create bank send tx command")
-
-	_, errBuf, err := s.upgradeManager.RunExec(ctx, exec)
-	s.Require().NoError(err, "failed to execute bank send tx")
-	// NOTE: The only message in the errBuf that is allowed is `gas estimate: ...`
-	gasEstimateMatch, err := regexp.MatchString(`^\s*gas estimate: \d+\s*$`, errBuf.String())
-	s.Require().NoError(err, "failed to match gas estimate message")
-	s.Require().True(
-		gasEstimateMatch,
-		"expected message in errBuf to be `gas estimate: ...`; got: %q\n",
-		errBuf.String(),
-	)
-
-	// Wait until the transaction has succeeded and is included in the chain
-	err = s.upgradeManager.WaitNBlocks(ctx, 2)
-	s.Require().NoError(err, "failed to wait for blocks")
-
-	balancePost, err := s.upgradeManager.GetBalance(ctx, s.upgradeParams.ChainID, receiver)
-	s.Require().NoError(err, "can't get balance of receiver account")
-
-	diff := balancePost.Sub(balancePre...)
-	s.Require().Equal(diff.String(), sentCoins.String(), "unexpected difference in bank balance")
-
 }
